GH-1677 [Bug] 京东服务器兼容 (#1681)
* 兼容京东服务器

* 添加sql

* 添加说明文档

* 格式化代码

* 提交错误

+++ spark-yun-backend/spark-yun-modules/src/main/java/com/isxcode/star/modules/monitor/entity/MonitorEntity.java
@@ -38,9 +38,9 @@ public class MonitorEntity {
 
     private String log;
 
-    private Long usedStorageSize;
+    private Double usedStorageSize;
 
-    private Long usedMemorySize;
+    private Double usedMemorySize;
 
     private Long networkIoReadSpeed;
